requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.23.0
trimesh>=3.20.0
pyrender>=0.1.45
opencv-python>=4.7.0
Pillow>=9.4.0
scikit-learn>=1.2.0
tqdm>=4.65.0
matplotlib>=3.6.0
pyglet>=2.0.0
pyopengl>=3.1.0